Replace os.unlink with delete_if_exists
Shouldn't care when doing unlink on our temp files
if they exist or not.  In fact this causes problems
when you do things like with tempfile/dir and happen
to try and unlink after it's already been removed.

This replaces these calls with the safer
common.fileutils.delete_if_exists which will
ignore the os exception of the object DNE.

Fixes bug: 1212502

Change-Id: Ica86c95f736411da486335aec5512e59247bfbc0