Fix errors in volume usage audit script
The snapshot_get_active_by_window DB query was returning snapshots
where we couldn't access the related volume object. This was due to
a) the volume not been loaded and the join failing due to detached
session object.
and b) the related volume was deleted so it didn't show return from
the join

Fixes bug: 1220140

Change-Id: I0de993ba0ca482fa9ab80d4d5cdb0ff5cc8e4d7a