This adds a README to brick
We need to let folks know where brick is
currently being maintained for the Havana
release and where to file defects.
Once this simple patch lands, we'll pull this into
the Nova patch.

(fix hacking violation in commit message for the 2nd
time on this patch. To prevent it from blocking the
merge queue on feature freeze day.)

PLEASE DON'T +A PATCHES UNTIL THEY PASS JENKINS CHECK

Change-Id: I5c4569c1c09751d53817ee9577581e530a3a8352