Fixes vol restore discard final bytes unzeroed
When doing a volume restore with Ceph backend to a non-rbd volume
that is greater in size that the backup volume, we write zeroes
the extraneous extents but we are not zeroing the final few bytes
in the case where the destination volume is non-rbd.

Also fixed up internationalisation on a number of log messages.

Change-Id: Iac1ff98a75e9164f8865c2f7d998454205e049d4
Fixes: bug 1226555