Fix type change in bootable setting of volume view
Change Idd47a0a8069ee905b81c7aae562b82767ad91930 fixed a db
query issue, by removing the image_metadata querie in the get/show
methods.

The problem however is that we just passed the setting from the volume
object directly in the summary view builder.  This results in changing
the type of the value from a string "true/false" to a bool True/False
which is no good.

This patch converts the object member back to a string like it used to
be so we don't break any compat or expectations in API behavior.

Fixes bug #1227837

Change-Id: I55c0e25134ae429802ddf313d6d26f1fad19ca94