Add a retry to create_iscsi_target for LVM
There's a bug in the tgt driver where under certain
conditions where a race condition is present that can
result in tgtadm attempting re-use a target ID.

Testing this it seems that a retry will address this nicely
and cause tgtd to grab a new target ID and things are ok.

This patch adds a simple check/retry mechanism around the
iscsi_target_create in the LVM driver to catch this issue and
deal with it.

There are a number of bugs that it appears can be attributed to
this, but this patch is associated with the most predominant
clear cut version.

Closes-Bug #1223469

Change-Id: I5126009f196adcafad55e73ff99a59262dd93dfe