Fix issues with failed lvremove
There are some race conditions that
can cause problems with lvremove commands.  In
most cases these seem to recover nicely just
with a simple retry of the lvremove.  Adding
a udev settle seems to elimate the rest of them.

This is a difficult issue to reproduce, and there's
a suspiscion that it relates to failed target
creeates.

The patch adds a catch on the lvremove failure,
followed by a udevadm settle and a retry of the
lvremove.  With the setup that I've been able
to reproduce this issue these changes have eliminated
any force delete failures.

The other option that had been proposed was using dmsetup remove
but there are concerns that this may cause problems.

Change-Id: I2a2b0d0f4fefd0daf9424ab96aaf87ba53ebc171
Closes-Bug: #1191960