Fix mandatory and optional args for create_volume
Calls to create_volume (both volume_rpcapi and scheduler_rpcapi) pass
mandatory args as optional and reverse. This patch fixes the calls to be
according to the function definition.

Also, fix the call to create_volume in the migration code where
allow_reschedule defaults to True, but we don't pass request_spec so it
prints a warning and disables it.

Change-Id: I9509f88866e0ae182f90ebdfca1fcf415ca60d19
Closes-Bug: #1230069