Fix Huawei HVS driver attaching volume error
If iSCSI initiator is not added to host, we will get https errors
for we find the initiator info by sending url with this initiator name.

This patch fixes the way of getting the initiator info.
First, get the all initiator info.
Then, find the one we need by name.

Closes-bug: #1230296
Change-Id: I92620374923fa136ee71fe6eb3af6e4c78a3d66b