Fixes ceph backup import errors
For some reason, using non-rbd volume driver whilst using
Ceph backup causes what looks to be a cyclic import error.
Ensuring that the backup.drivers.ceph module is imported
into volume.drivers.rbd ONLY if it is needed i.e. if both
drivers are in use, seems to remedy this.

Change-Id: Ieae0c3451cfb1750eca8cfcde5e0be9e7eb028d0
Fixes: bug 1231412