Add debug logging for targets
Seems we may be failing target create silently,
the situation appears to be that the actual device
lun is not created.  This has been appearing in the
gates on create from snapshot.

This change adds some logging that will help us determine
if it's the actual create, or if it's the attempted update
while the dd process is possibly holding the volume in a busy
state causing the target update to add lun 1 to fail.

Change-Id: If5d4219dedae4649982fa9c13f131bf5f38fb7c4