Dont retry if target creation succeeds
The target creation retry loop was not breaking when
the target was succesfully created.  This is "sometimes" ok,
however in other cases this will actually cause a failure in
the second create and the backing Lun will be deleted and not
created again succesfully due most likely to a busy status on the
target.

Add a break in the try block.

Change-Id: I875f6302868651b5b42d4796cd2714fba35e525e
Closes-Bug: #1226337