Revert "Brick connector fix for NFS drivers"
This reverts commit b6b9df2940ac07f715f759f4c315b2cf088c2320

Passing in driver=self in the base driver connector builder breaks all of the other drivers and isn't the *correct* driver to be passing in.

Simply removing this fixes the functionality, however breaks your tests and I'm not sure of the impact on what you were attempting so I'm reverting.

Change-Id: I0af85a2bc491cc0e23aa6af1f67c2dfed1c7081f