Handle GlanceMetadataNotFound Exceptions
Change number I859bb2550267c6a5142e85ec89f1c7ff885588ac
fixed a bug where we weren't copying the bootable properties
of a volume or snapshot on create from snap or volume.

This didn't however handle the case of non-bootable volumes/snaps.

The result is that when the db call to get the image metadata raises
a NotFound exception we don't handle it and crash, causing the create
operation to fail.

This adds a try/except block around the copy volume/snap
glance_metadata get calls and ignores the exception.

Change-Id: Ibad730e7886b082d34717cd01b4ba31941e7594a
Closes-Bug: 1246004